backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Gallery backend"
authors = ["Christian Franchin dos Santos <christianfranchin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.3"
Flask-HTTPAuth = "^4.5.0"
autopep8 = "^1.6.0"
PyJWT = "^2.3.0"
pymongo = "^4.0.1"

[tool.poetry.dev-dependencies]
pylama = "^8.3.7"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
